requirements.txt
# ── Microsoft Agent Framework ────────────────────────────────────────────
# SDK in preview — pin exact versions to avoid breaking renames
agent-framework-azure-ai==1.0.0rc3
agent-framework-core==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16

# ── Azure identity + clients ─────────────────────────────────────────────
azure-identity>=1.19.0
azure-cosmos>=4.7.0
azure-ai-projects>=1.0.0b11

# ── Streamlit frontend ───────────────────────────────────────────────────
streamlit>=1.42.0

# ── Auth ─────────────────────────────────────────────────────────────────
msal>=1.28.0

# ── Utilities ────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
pyyaml>=6.0.2
pypdf>=4.0.0
python-docx>=1.1.0
requests>=2.31.0
